smart_validation_final/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smart_validation"
version = "0.1.1"
description = "Smart Validation Tool for SDTM/ADaM/TFL"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas",
    "openpyxl",
    "pyyaml",
    "typer[all]",
    "rich",
    "streamlit",
]

[project.scripts]
smartvalidator = "smart_validation.cli:main"
smartvalidator-gui = "smart_validation.app:main"
